refactor: corrige tratamento de erro no fidelity

+++ ecommerce/src/main/java/br/ufrn/imd/ecommerce/dto/FidelityRequestDto.java
@@ -1,4 +1,4 @@
 package br.ufrn.imd.ecommerce.dto;
 
-public record FidelityRequestDto(String idUser, Double bonus) {
+public record FidelityRequestDto(String idUser, Integer bonus) {
 }
